+++ packages/ui/src/lib/components/discord/MessageEmbedFooter.tsx
@@ -1,0 +1,7 @@
+export interface IDiscordMessageEmbedFooter {
+	content: string;
+}
+
+export function DiscordMessageEmbedFooter({ content }: IDiscordMessageEmbedFooter) {
+	return <div className="mt-2 text-xs">{content}</div>;
+}