+++ packages/ws/src/utils/constants.ts
@@ -1,5 +1,3 @@
-import { readFileSync } from 'node:fs';
-import { join } from 'node:path';
 import process from 'node:process';
 import { Collection } from '@discordjs/collection';
 import { lazy } from '@discordjs/util';
@@ -20,10 +18,7 @@ export enum CompressionMethod {
 	ZlibStream = 'zlib-stream',
 }
 
-const packageJson = readFileSync(join(__dirname, '..', '..', 'package.json'), 'utf8');
-const Package = JSON.parse(packageJson);
-
-export const DefaultDeviceProperty = `@discordjs/ws ${Package.version}`;
+export const DefaultDeviceProperty = `@discordjs/ws [VI]{{inject}}[/VI]`;
 
 const getDefaultSessionStore = lazy(() => new Collection<number, SessionInfo | null>());